setup local deployment process, let ci deploy locally and run smoke tests, adjust readme

+++ make.sh
@@ -1,6 +1,8 @@
 #!/bin/bash -e
 
-mode=$1
+local=$1
+mode=$2
+
 if [[ -f config.env ]]
 then
     export $(cat config.env | sed 's/#.*//g' | xargs)
@@ -8,8 +10,18 @@ fi
 
 build_and_deploy_backend() {
 
+    IS_LOCAL=$1
     cwd=$(pwd)
     cd src/backend/layer
+    if [[ ! -f ./bin/phantomjs ]]; then
+        mkdir -p tmp
+        cd tmp
+        wget https://bitbucket.org/ariya/phantomjs/downloads/phantomjs-2.1.1-linux-x86_64.tar.bz2
+        tar -xf ./phantomjs-2.1.1-linux-x86_64.tar.bz2 -C ./
+        mv phantomjs-2.1.1-linux-x86_64/bin/phantomjs ../bin/phantomjs
+        cd ..
+        rm -rf tmp
+    fi
     zip -r layer.zip bin
     cd $cwd
     awk -v key=$GUESSR_MAPS_API_KEY '{gsub("<%= MAPS_KEY %>", key, $0); print}' src/backend/layer/bin/template.html > src/backend/layer/bin/index.html
@@ -27,22 +39,29 @@ build_and_deploy_backend() {
     echo "------------------------"
     echo "Done!"
     echo ""
-    echo "------------------------"
-    echo "Deploying Backend..."
-    echo "------------------------"
-    if [[ ${mode} = "guided" ]]; then
-        rm samconfig.toml
-        sam deploy --guided
-    else
-        sam deploy
-    fi
 
-    if (( $? != 0 )); then
-        exit
+    if [[ -z "$IS_LOCAL" ]]
+    then
+        echo "------------------------"
+        echo "Deploying Backend..."
+        echo "------------------------"
+        if [[ ${mode} = "guided" ]]; then
+            rm samconfig.toml
+            sam deploy --guided
+        else
+            sam deploy
+        fi
+        if (( $? != 0 )); then
+            exit
+        fi
+    else
+        echo "============================================================"
+        echo "====                                                    ===="
+        echo "==== DON'T FORGET TO START THE DYNAMO DB DOCKER COMPOSE ===="
+        echo "====                                                    ===="
+        echo "============================================================"
+        sam local start-api --docker-network sam-local-network
     fi
-    echo "------------------------"
-    echo "Done!"
-    echo ""
 }
 
 build_frontend() {
@@ -50,16 +69,18 @@ build_frontend() {
     echo "Building Frontend..."
     echo "------------------------"
 
-    RES=$(aws cloudformation describe-stack-resources --stack-name georgguessr)
-    API_KEY_ID=$(echo ${RES} | jq -r '.StackResources[]  | select(.ResourceType == "AWS::ApiGateway::ApiKey") | .PhysicalResourceId')
-    API_KEY_VALUE=$(aws apigateway get-api-key --include-value --api-key ${API_KEY_ID} | jq -r '.value')
-    API_ID=$(echo ${RES} | jq -r '.StackResources[]  | select(.ResourceType == "AWS::ApiGateway::RestApi") | .PhysicalResourceId')
-    API_ENDPOINT="https://$API_ID.execute-api.eu-west-1.amazonaws.com/Prod"
-
-
-    echo "---"
-    echo "$API_KEY_VALUE"
-    echo "---"
+    IS_LOCAL=$1
+    if [[ -z "$IS_LOCAL" ]]
+    then
+        RES=$(aws cloudformation describe-stack-resources --stack-name georgguessr)
+        API_KEY_ID=$(echo ${RES} | jq -r '.StackResources[]  | select(.ResourceType == "AWS::ApiGateway::ApiKey") | .PhysicalResourceId')
+        API_KEY_VALUE=$(aws apigateway get-api-key --include-value --api-key ${API_KEY_ID} | jq -r '.value')
+        API_ID=$(echo ${RES} | jq -r '.StackResources[]  | select(.ResourceType == "AWS::ApiGateway::RestApi") | .PhysicalResourceId')
+        API_ENDPOINT="https://$API_ID.execute-api.eu-west-1.amazonaws.com/Prod"
+    else
+        API_KEY_VALUE=""
+        API_ENDPOINT="http://127.0.0.1:3000"
+    fi
 
     cd src/frontend
     npm install
@@ -72,13 +93,12 @@ build_frontend() {
 }
 
 ###################################
-
-if [[ ${mode} != "frontend" ]]; then
+if [[ ${local} != "local" ]]; then
+    echo "deploying remote"
     build_and_deploy_backend
-fi
-
-if [[ ${mode} != "backend" ]]; then
     build_frontend
+else
+    echo "running locally"
+    build_frontend "$local"
+    build_and_deploy_backend "$local"
 fi
-
-echo "Run the frontend local with ' http-server dist '"
